refactored resources code
